+++ selene/support/shared/config.py
@@ -159,6 +159,7 @@ class SharedConfig(Config):
         base_url: str = '',
         set_value_by_js: bool = False,
         type_by_js: bool = False,
+        click_by_js: bool = False,
         wait_for_no_overlap_found_by_js: bool = False,
         window_width: Optional[int] = None,
         window_height: Optional[int] = None,
@@ -231,6 +232,7 @@ class SharedConfig(Config):
             base_url=base_url,
             set_value_by_js=set_value_by_js,
             type_by_js=type_by_js,
+            click_by_js=click_by_js,
             wait_for_no_overlap_found_by_js=wait_for_no_overlap_found_by_js,
             window_width=window_width,
             window_height=window_height,
@@ -423,6 +425,10 @@ PageSource: file://{path}'''
     @Config.type_by_js.setter
     def type_by_js(self, value: bool):
         self._type_by_js = value
+
+    @Config.click_by_js.setter
+    def click_by_js(self, value: bool):
+        self._click_by_js = value
 
     @Config.wait_for_no_overlap_found_by_js.setter
     def wait_for_no_overlap_found_by_js(self, value: bool):